Component `LLConfig`
