+++ LLDebugTool/Core/Component/Magnifier/UserInterface/LLMagnifierViewController.m
@@ -26,8 +26,9 @@
 #import "LLMagnifierInfoView.h"
 #import "LLInternalMacros.h"
 #import "LLMagnifierView.h"
-#import "LLConfig.h"
 #import "LLConst.h"
+
+#import "LLConfig+Magnifier.h"
 
 @interface LLMagnifierViewController ()<LLMagnifierViewDelegate, LLInfoViewDelegate>
