+++ LLDebugTool/Core/Component/Network/Function/LLConfig+Network.m
@@ -1,5 +1,5 @@
 //
-//  LLLogHelper.h
+//  LLConfig+Network.m
 //
 //  Copyright (c) 2018 LLDebugTool Software Foundation (https://github.com/HDB-Li/LLDebugTool)
 //
@@ -21,46 +21,34 @@
 //  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 //  SOFTWARE.
 
-#import <Foundation/Foundation.h>
+#import "LLConfig+Network.h"
 
-#import "LLConfig+Log.h"
+#import "NSObject+LL_Runtime.h"
 
-NS_ASSUME_NONNULL_BEGIN
+@implementation LLConfig (Network)
 
-/**
- Quick print and save log.
- */
-@interface LLLogHelper : NSObject
+- (void)setObserverdHosts:(NSArray<NSString *> *)observerdHosts {
+    objc_setAssociatedObject(self, @selector(observerdHosts), observerdHosts, OBJC_ASSOCIATION_RETAIN_NONATOMIC);
+}
 
-/**
- Singleton to control enable.
- 
- @return Singleton
- */
-+ (instancetype)shared;
+- (NSArray<NSString *> *)observerdHosts {
+    return objc_getAssociatedObject(self, _cmd);
+}
 
-/**
- Set enable to save log model.
- */
-@property (nonatomic, assign, getter=isEnabled) BOOL enable;
+- (void)setIgnoredHosts:(NSArray<NSString *> *)ignoredHosts {
+    objc_setAssociatedObject(self, @selector(ignoredHosts), ignoredHosts, OBJC_ASSOCIATION_RETAIN_NONATOMIC);
+}
 
-/**
- Return log levels string.
- */
-+ (NSArray <NSString *>*)levelsDescription;
+- (NSArray<NSString *> *)ignoredHosts {
+    return objc_getAssociatedObject(self, _cmd);
+}
 
-/**
- Print and save a log model with infos.
+- (void)setObserverWebView:(BOOL)observerWebView {
+    objc_setAssociatedObject(self, @selector(observerWebView), @(observerWebView), OBJC_ASSOCIATION_RETAIN_NONATOMIC);
+}
 
- @param file File name.
- @param function Function name.
- @param lineNo Line number.
- @param level Log level.
- @param onEvent Event,can filter by this.
- @param message Message.
- */
-- (void)logInFile:(NSString *_Nullable)file function:(NSString *_Nullable)function lineNo:(NSInteger)lineNo level:(LLConfigLogLevel)level onEvent:(NSString *_Nullable)onEvent message:(NSString *_Nullable)message;
+- (BOOL)observerWebView {
+    return [objc_getAssociatedObject(self, _cmd) boolValue];
+}
 
 @end
-
-NS_ASSUME_NONNULL_END
