+++ LLDebugTool/DebugTool/LLConfig.h
@@ -42,21 +42,6 @@ typedef NS_ENUM(NSUInteger, LLConfigColorStyle) {
 };
 
 /**
- Log Level. It can be used for filter.
- 
- - LLConfigLogLevelDefault: Use to save message or note.
- - LLConfigLogLevelAlert: Use to save alert message.
- - LLConfigLogLevelWarning: Use to save warning message.
- - LLConfigLogLevelError: Use to save error message.
- */
-typedef NS_ENUM(NSUInteger, LLConfigLogLevel) {
-    LLConfigLogLevelDefault,
-    LLConfigLogLevelAlert,
-    LLConfigLogLevelWarning,
-    LLConfigLogLevelError,
-};
-
-/**
  Window style. Decide how the Window displays.
  
  - LLConfigEntryWindowStyleBall: Show as a ball. Moveable and clickable.
@@ -80,23 +65,6 @@ typedef NS_ENUM(NSUInteger, LLConfigEntryWindowStyle) {
     LLConfigEntryWindowStylePowerBar = 5,
 #endif
     LLConfigEntryWindowStyleSuspensionBall NS_ENUM_DEPRECATED_IOS(2_0, 8_0, "Use LLConfigEntryWindowStyleBall") = 0,
-};
-
-/**
- Log style for [LLDebugTool logInFile...]. Customize the log you want.
- 
- - LLConfigLogDetail: Show all detail info. Contain event, file, line, func, date and desc.
- - LLConfigLogFileFuncDesc : Show with event, file, func and desc.
- - LLConfigLogFileDesc : Show with event, file and desc.
- - LLConfigLogNormal: Show as system NSLog
- - LLConfigLogNone: Don't show anything.
- */
-typedef NS_ENUM(NSUInteger, LLConfigLogStyle) {
-    LLConfigLogDetail,
-    LLConfigLogFileFuncDesc,
-    LLConfigLogFileDesc,
-    LLConfigLogNormal,
-    LLConfigLogNone,
 };
 
 /**
@@ -146,6 +114,8 @@ typedef NS_OPTIONS(NSUInteger, LLConfigAvailableFeature) {
  - LLDebugToolActionMagnifier: Magnifier function.
  - LLDebugToolActionRuler: Ruler function.
  - LLDebugToolActionWidgetBorder: Widget border function.
+ - LLDebugToolActionHtml: Html function.
+ - LLDebugToolActionLocation: Mock location function.
  */
 typedef NS_ENUM(NSUInteger, LLDebugToolAction) {
     LLDebugToolActionFunction,
@@ -237,12 +207,6 @@ NS_ASSUME_NONNULL_BEGIN
  */
 - (void)configStatusBarStyle:(UIStatusBarStyle)statusBarStyle;
 
-#pragma mark - Log
-/**
- Customize the log style. Default is LLConfigLogDetail.
- */
-@property (nonatomic, assign) LLConfigLogStyle logStyle;
-
 #pragma mark - Date Formatter
 /**
  Date Format Style. Use to recording time when create model. Default is "yyyy-MM-dd HH:mm:ss".
@@ -250,64 +214,11 @@ NS_ASSUME_NONNULL_BEGIN
  */
 @property (nonatomic, copy) NSString *dateFormatter;
 
-#pragma mark - Magnifier Window
-/**
- Magnifier window zoom level, number of pixels per color, default is kLLMagnifierWindowZoomLevel.
- */
-@property (nonatomic, assign) NSInteger magnifierZoomLevel;
-
-/**
- Number of rows per magnifier window, default is kLLMagnifierWindowSize.
- */
-@property (nonatomic, assign) NSInteger magnifierSize;
-
 #pragma mark - User Identity
 /**
  Tag user name is used to create the crash/network/log model.
  */
 @property (nonatomic, copy, nullable) NSString *userIdentity;
-
-#pragma mark - Network
-/**
- Observer network in hosts, ignore others.
- */
-@property (nonatomic, strong, nullable) NSArray <NSString *>*observerdHosts;
-
-/**
- Ignored hosts, low level than observerdHosts.
- */
-@property (nonatomic, strong, nullable) NSArray <NSString *>*ignoredHosts;
-
-/**
- Whether observer webView request.
- */
-@property (nonatomic, assign) BOOL observerWebView;
-
-#pragma mark - Hierarchy
-
-/**
- Hierarchy function ignore private class or not.
- */
-@property (nonatomic, assign, getter=isHierarchyIgnorePrivateClass) BOOL hierarchyIgnorePrivateClass;
-
-#pragma mark - Widget Border
-
-/**
-Whether show widget border. Default is NO.
-*/
-@property (nonatomic, assign, getter=isShowWidgetBorder) BOOL showWidgetBorder;
-
-#pragma mark - Html5
-
-/**
- Default html5 url string used in Html function. must has prefix with http:// or https://
- */
-@property (nonatomic, copy, nullable) NSString *defaultHtmlUrl;
-
-/**
- Custom view controller used in html function. you can use your custom viewController to dynamic debug your web view. must comply with `LLComponentDelegate`. ViewController must set background color.
- */
-@property (nonatomic, copy, nullable) UIViewController *(^htmlViewControllerProvider)(NSString * _Nullable url);
 
 #pragma mark - Location
 /**
